refactor ObjSymbol to use 32 bit addressing/sizing

+++ src/analysis/mod.rs
@@ -148,9 +148,7 @@ fn get_jump_table_entries(
                 .kind_at_section_address(addr.section, addr.address, ObjSymbolKind::Object)
                 .ok()
                 .flatten()
-                .and_then(
-                    |(_, s)| if s.size_known { NonZeroU32::new(s.size as u32) } else { None },
-                );
+                .and_then(|(_, s)| if s.size_known { NonZeroU32::new(s.size) } else { None });
             if let Some(size) = known_size.or(size).map(|n| n.get()) {
                 log::trace!(
                     "Located jump table @ {:#010X} with entry count {} (from {:#010X})",
@@ -228,9 +226,7 @@ fn get_jump_table_entries(
                 .kind_at_section_address(addr.section, addr.address, ObjSymbolKind::Object)
                 .ok()
                 .flatten()
-                .and_then(
-                    |(_, s)| if s.size_known { NonZeroU32::new(s.size as u32) } else { None },
-                );
+                .and_then(|(_, s)| if s.size_known { NonZeroU32::new(s.size) } else { None });
             if let Some(size) = known_size.or(size).map(|n| n.get()) {
                 log::trace!(
                     "Located jump table @ {:#010X} with entry count {} (from {:#010X})",
